feat(report): refactor code
