+++ src/redux/actions/incident.js
@@ -1,5 +1,6 @@
+/* eslint-disable consistent-return */
 import axios from 'axios';
-import { handleMessages } from '../../utils/helpers';
+import { handleMessages, uploadMedia } from '../../utils/helpers';
 
 export const REPORT_INCIDENT_REQUEST = 'REPORT_INCIDENT_REQUEST';
 export const REPORT_INCIDENT_SUCCESS = 'REPORT_INCIDENT_SUCCESS';
@@ -19,41 +20,29 @@ export const reportIncidentFailureAction = payload => ({
 
 export const reportIncident = data => async dispatch => {
   dispatch(reportIncidentRequestAction());
+  const invalidIncidentTypeMsg = 'Please select a type of incident';
 
   try {
-    const { Images, Videos } = data;
-    const uploads = [...Images, ...Videos].map(file => {
-      const formData = new FormData();
-      formData.append('file', file);
-      formData.append('upload_preset', process.env.UPLOAD_PRESET);
-      formData.append('timestamp', Date.now());
+    if (!data.type) throw new Error(invalidIncidentTypeMsg);
 
-      return axios.post(
-        `https://api.cloudinary.com/v1_1/${process.env.CLOUD_NAME}/upload`,
-        formData
-      );
-    });
-
-    const uploaded = uploads.length ? await axios.all(uploads) : [];
-
-    const media = uploaded.map(m => ({
-      url: m.data.secure_url,
-      type: m.data.resource_type,
-    }));
-
-    data.Images = media.filter(m => m.type === 'image').map(img => img.url);
-    data.Videos = media.filter(m => m.type === 'video').map(vid => vid.url);
     const token = localStorage.getItem('token');
+
+    await uploadMedia(data);
+
     const res = await axios.post(
       `${process.env.API_BASE_URL}/${data.type}s`,
       { ...data },
       { headers: { 'access-token': token } }
     );
     const { message, ...incident } = res.data.data[0];
+
     handleMessages([message], 'success');
     dispatch(reportIncidentSuccessAction(incident));
   } catch (err) {
-    let messages = ['An error occurred'];
+    let messages =
+      err.message === invalidIncidentTypeMsg
+        ? [err.message]
+        : ['An error occurred'];
 
     if (err.response) {
       const { error, errors } = err.response.data;
